+++ src/GafferSceneUI/LightPositionTool.cpp
@@ -397,6 +397,11 @@ class DistanceHandle : public Handle
 			m_requiresPivot = requiresPivot;
 		}
 
+		bool getRequiresPivot() const
+		{
+			return m_requiresPivot;
+		}
+
 	protected :
 
 		void renderHandle( const Style *style, Style::State state ) const override
@@ -635,18 +640,27 @@ void LightPositionTool::positionHighlight(
 	const float targetDistance
 )
 {
+	const V3f reflectionRay = reflect( ( viewpoint - highlightTarget ), normal ).normalize();
+	positionAlongNormal( highlightTarget, reflectionRay, targetDistance );
+}
+
+void LightPositionTool::positionAlongNormal(
+	const V3f &target,
+	const V3f &normal,
+	const float distance
+)
+{
 	if( !m_distanceHandle->enabled() || selection().empty() )
 	{
 		return;
 	}
 	const Selection &s = selection().back();
 
-	const V3f reflectionRay = reflect( ( viewpoint - highlightTarget ), normal ).normalize();
-	const V3f newP = highlightTarget + reflectionRay * targetDistance;
+	const V3f newP = target + normal * distance;
 
 	Context::Scope scopedContext( s.context() );
 
-	const M44f orientationMatrix = sourceOrientation( s, newP, highlightTarget );
+	const M44f orientationMatrix = sourceOrientation ( s, newP, target );
 
 	const M44f localTransform = s.scene()->transform( s.path() );
 	translateAndOrient( s, localTransform, newP, orientationMatrix );
@@ -811,7 +825,8 @@ RunTimeTypedPtr LightPositionTool::sceneGadgetDragBegin( Gadget *gadget, const D
 	{
 		return nullptr;
 	}
-	if( getTargetMode() == TargetMode::Pivot && modePlug()->getValue() == (int)Mode::Highlight )
+	const auto distanceHandle = static_cast<DistanceHandle *>( m_distanceHandle.get() );
+	if( getTargetMode() == TargetMode::Pivot && !distanceHandle->getRequiresPivot() )
 	{
 		return nullptr;
 	}
@@ -939,7 +954,8 @@ bool LightPositionTool::buttonPress( const ButtonEvent &event )
 		return true;
 	}
 
-	if( getTargetMode() == TargetMode::Pivot && modePlug()->getValue() == (int)Mode::Highlight )
+	const auto distanceHandle = static_cast<DistanceHandle *>( m_distanceHandle.get() );
+	if( getTargetMode() == TargetMode::Pivot && !distanceHandle->getRequiresPivot() )
 	{
 		return true;
 	}
@@ -995,7 +1011,7 @@ bool LightPositionTool::placeTarget( const LineSegment3f &eventLine )
 	}
 	else if( getTargetMode() == TargetMode::Target )
 	{
-		if( modePlug()->getValue() == (int)Mode::Highlight && !distanceHandle->getTarget() )
+		if( !distanceHandle->getRequiresPivot() && !distanceHandle->getTarget() )
 		{
 			setPivotDistance(
 				(
@@ -1037,6 +1053,26 @@ bool LightPositionTool::placeTarget( const LineSegment3f &eventLine )
 			positionHighlight(
 				distanceHandle->getTarget().value() * sceneToTransformSpaceInverse,
 				cameraTransform.translation(),
+				worldNormal,
+				distanceHandle->getPivotDistance().value()
+			);
+		}
+	}
+	else
+	{
+		if( !distanceHandle->getTarget() )
+		{
+			return false;
+		}
+
+		std::optional<V3f> sceneGadgetNormal = sceneGadget->normalAt( eventLine );
+		if( sceneGadgetNormal )
+		{
+			V3f worldNormal;
+			sceneGadget->fullTransform().inverse().transpose().multDirMatrix( sceneGadgetNormal.value(), worldNormal );
+
+			positionAlongNormal(
+				distanceHandle->getTarget().value() * sceneToTransformSpaceInverse,
 				worldNormal,
 				distanceHandle->getPivotDistance().value()
 			);